Create 2019-08-26-rameel-khan-site-launched.md
Jekyll is very strict with how these files are named, so pay attention. It must follow the convention YYYY-MM-DD-title-of-my-post.md
This file name gets translated into the permalink for the blog post.
Note the file extension .md stands for Markdown, and the Markdown syntax used inside the file gets converted to HTML by Jekyll.